Refactor monetary values to use integer cents representation across the application
- Updated price formatting in ItemsSection and ReceiptWidgets to use MoneyFormatter.
- Changed monetary fields in CustomerReceiptWidget, ZReportReceiptWidget, and CloseShiftDialog to use integers instead of doubles.
- Refactored Shift model and repository to handle monetary values as integers (cents).
- Adjusted OpenShift and CloseShift use cases to accept and return integer values for cash amounts.
- Enhanced settings screen to convert delivery fee to cents.
- Ensured all relevant calculations and displays reflect the new integer representation for monetary values.

+++ lib/features/expenses/presentation/bloc/expenses_bloc.dart
@@ -27,7 +27,7 @@ class ExpensesBloc extends Bloc<ExpensesEvent, ExpensesState> {
     result.fold(
       (failure) => emit(ExpensesError(failure.message)),
       (expenses) {
-        final double totalExpenses = expenses.fold(0.0, (sum, item) => sum + item.amount);
+        final int totalExpenses = expenses.fold(0, (sum, item) => sum + item.amount);
         emit(ExpensesLoaded(expenses, totalExpenses));
       },
     );
