+++ lib/features/shift/domain/repositories/shift_repository.dart
@@ -4,6 +4,6 @@ import '../entities/shift.dart';
 
 abstract class ShiftRepository {
   Future<Either<Failure, Shift?>> checkActiveShift();
-  Future<Either<Failure, Shift>> openShift({required double startingCash, required int cashierId});
+  Future<Either<Failure, Shift>> openShift({required int startingCash, required int cashierId});
   Future<Either<Failure, Shift>> closeShift({required int shiftId, required double actualCash});
 }
